+++ src/api/consumer/lib/test.js
@@ -8,6 +8,7 @@ export async function test(id: string, verbose: boolean = true): Promise<Bit> {
   const consumer: Consumer = await loadConsumer();
   const idParsed = BitId.parse(id);
   const component = await consumer.loadComponent(idParsed);
+  if (!component.testerId) return { component, missingTester:true }
   const result = await consumer.runComponentSpecs(idParsed, verbose);
   return  { specs: result, component };
 }
@@ -17,8 +18,12 @@ export async function testAll(verbose: boolean = true): Promise<Bit> {
   const componentsList = new ComponentsList(consumer);
   const newAndModifiedComponents = await componentsList.newAndModifiedComponents();
   const specsResults = newAndModifiedComponents.map(async (component) => {
-    const result = await component.runSpecs({ scope: consumer.scope, consumer, verbose });
-    return { specs: result, component };
+    if (!component.testerId) {
+      return { component, missingTester:true }
+    } else {
+      const result = await component.runSpecs({scope: consumer.scope, consumer, verbose});
+      return { specs: result, component };
+    }
   });
 
   return Promise.all(specsResults);